refactor use of fetch api by frontend code
