+++ src/packages/frontend/client/purchases.ts
@@ -3,29 +3,7 @@ Functions for interfacing with the purchases functionality.
 */
 // import type { WebappClient } from "./client";
 
-// TODO: refactor with components/run-button/api.ts
-import { join } from "path";
-import { appBasePath } from "@cocalc/frontend/customize/app-base-path";
-
-export default async function api(endpoint: string, args?: object) {
-  const url = join(appBasePath, "api/v2", endpoint);
-  const resp = await (
-    await fetch(url, {
-      method: args != null ? "POST" : "GET",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      ...(args != null ? { body: JSON.stringify(args) } : undefined),
-    })
-  ).json();
-  if (resp == null) {
-    throw Error("timeout -- please try again");
-  }
-  if (resp.error) {
-    throw Error(resp.error);
-  }
-  return resp;
-}
+import api from "./api";
 
 export class PurchasesClient {
   //private client: WebappClient;
